+++ src/main/java/com/grmkris/btcrbtcswapper/bitfinex/BitfinexWatcher.java
@@ -19,6 +19,7 @@ import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
@@ -84,7 +85,7 @@ public class BitfinexWatcher {
                         result = bitfinexHandler.convertBTCToLightning(amount.toString());
                         log.info("Converted BTC to Lightning, {}", result);
                         TimeUnit.SECONDS.sleep(60);
-                        String invoice = lndHandler.getLightningInvoice(amount.multiply(BigDecimal.valueOf(100000000)));
+                        String invoice = lndHandler.getLightningInvoice(amount.multiply(BigDecimal.valueOf(100000000)).toBigInteger());
                         result = bitfinexHandler.withdrawLightning(invoice);
                         log.info("Withdrew Lightning, {}", result);
                         log.info("Returning balancing status back to IDLE");
